starlark-test crate
* No need to generate test files (`build.rs`) when depending on starlark crate
* No need to compile `testutil` for each test
* `benches` and `tests` may share code (but not yet)
