+++ starlark-test/tests/go-testcases/assign.sky
