+++ starlark-test/tests/go-testcases/dict.sky
