+++ starlark-test/tests/go-testcases/function.sky
