+++ starlark-test/tests/go-testcases/list.sky
