+++ starlark-test/tests/go-testcases/struct.sky
