+++ starlark-test/tests/go-testcases/tuple.sky
