+++ starlark-test/tests/java-testcases/list_slices.sky
