+++ starlark-test/tests/java-testcases/min_max.sky
