+++ starlark-test/tests/java-testcases/string_format.sky
