+++ starlark-test/tests/java-testcases/string_splitlines.sky
