+++ starlark-test/tests/java-testcases/string_test_characters.sky
