+++ starlark-test/tests/rust-testcases/mutation_during_iteration.sky
