+++ starlark-test/tests/rust-testcases/struct.sky
